Move algorithm files and utils in algos directory
